Replace BaseSystem with BaseTable
