+++ OpenKh.Kh2/BaseTable.cs
@@ -42,4 +42,41 @@ namespace OpenKh.Kh2
 
         IEnumerator IEnumerable.GetEnumerator() => Items.GetEnumerator();
     }
+
+    public class BaseShortTable<T> : IEnumerable<T>
+        where T : class
+    {
+        [Data] public short Id { get; set; }
+        [Data] public short Count { get; set; }
+        public List<T> Items { get; set; }
+
+        public static BaseShortTable<T> Read(Stream stream)
+        {
+            var baseTable = BinaryMapping.ReadObject<BaseShortTable<T>>(stream);
+            baseTable.Items = Enumerable.Range(0, baseTable.Count)
+                .Select(_ => BinaryMapping.ReadObject<T>(stream))
+                .ToList();
+            return baseTable;
+        }
+
+        public static void Write(Stream stream, int id, List<T> items) =>
+            new BaseShortTable<T>()
+            {
+                Id = (short)id,
+                Items = items
+            }.Write(stream);
+
+
+        public void Write(Stream stream)
+        {
+            Count = (short)Items.Count;
+            BinaryMapping.WriteObject(stream, this);
+            foreach (var item in Items)
+                BinaryMapping.WriteObject(stream, item);
+        }
+
+        public IEnumerator<T> GetEnumerator() => Items.GetEnumerator();
+
+        IEnumerator IEnumerable.GetEnumerator() => Items.GetEnumerator();
+    }
 }
