+++ OpenKh.Kh2/System/Trsr.cs
@@ -24,9 +24,9 @@ namespace OpenKh.Kh2.System
         [Data] public short EventId { get; set; }
         [Data] public short OverallChestIndex { get; set; }
 
-        public static List<Trsr> Read(Stream stream) => BaseSystem<Trsr>.Read(stream).Items;
+        public static List<Trsr> Read(Stream stream) => BaseShortTable<Trsr>.Read(stream).Items;
 
-        public static void Write(Stream stream, IEnumerable<Trsr> items) => new BaseSystem<Trsr>
+        public static void Write(Stream stream, IEnumerable<Trsr> items) => new BaseShortTable<Trsr>
         {
             Id = MagicHeader,
             Items = items.ToList()
