+++ OpenKh.Tests/kh2/TrsrTests.cs
@@ -1,6 +1,7 @@
 ﻿using Xunit;
 using OpenKh.Common;
 using OpenKh.Kh2.System;
+using OpenKh.Kh2;
 
 namespace OpenKh.Tests.kh2
 {
@@ -9,7 +10,7 @@ namespace OpenKh.Tests.kh2
         [Fact]
         public void CheckNewTrsr() => Common.FileOpenRead(@"kh2/res/trsr.bin", x => x.Using(stream =>
         {
-            var table = BaseSystem<Trsr>.Read(stream);
+            var table = BaseTable<Trsr>.Read(stream);
             Assert.Equal(0x1AE, table.Count);
         }));
     }
